Fix backward advancePage w/ padding (#110)
* Fix backward advancePage w/ padding

If you use a padding on the carousel items, the currentPage variable becomes a float number which goes < 0. This made impossibile to go back to the first item (it was in stuck with the 2nd slide). Using Math.ceil on the page number fixed the issue.

* docs(standard): Update README.md

- add table with configuration options (property, type, default, description)
- add some properties in demo HTML markup to show how configuration is done

* Update README.md

Changed CamelCase to kebab-case in sample in ReadMe.md (because its not working).
https://vuejs.org/v2/guide/components.html#camelCase-vs-kebab-case

* unused duplicated methods

* remove unusded comment