#97 feat(props): add `adjustableHeight` prop (#292)
* #97 feat(props): add `adjustableHeight` prop

Allow the height of the carousel to change based on its current content.

* fix(example): Set fixed width slide content

* fix(perf): only watch additional mutation events for adjustableHeight

* fix(lint): Missing semi-colon

Fix listing issue

* fix(bug): adjustableHeight - Incorrect initial height

The initial `currentHeight` prop of the carousel was not calculating correctly because the CSS was constraining the width of the `slide`. This is only noticeable when slide content is slow to load. The additional CSS ensures the slide width is correct when calculating the initial carousel height.

* fix(bug): Add missing return statement